Restructure Part 1
The FW has gotten bloated, and needs restructuring.

First, by using a shared context variable, we can reduce the amount of
shared pointers to variables.

Second, we should not be defining any variables in header files, only
declarations and preprocessor directives.

Third, we should start removing unnecessary or unclear comments.
The code should be straight forward enough to read that comments are
largely not needed, except for complex sections.

Fourth, we should be using TickType_t to time events instead of
millis().